+++ HyperStoreServiceAPP/Models/Product.cs
@@ -13,7 +13,7 @@ namespace HyperStoreService.Models
         [Required]
         public string Code { get; set; }
         [Required]
-        public float? DisplayPrice { get; set; }
+        public decimal? DisplayPrice { get; set; }
         public float DiscountPer { get; set; }
         [Required]
         public string Name { get; set; }
